+++ tests/test_strip.py
@@ -17,11 +17,14 @@
 """
 
 import inspect
+import time
 
 import numpy as np
 
 from aptapy.plotting import plt
-from aptapy.strip import StripChart
+from aptapy.strip import EpochStripChart, StripChart
+
+_RNG = np.random.default_rng(313)
 
 
 def test_strip_chart_seconds():
@@ -36,15 +39,17 @@ def test_strip_chart_seconds():
     plt.legend()
 
 
-def test_strip_chart_datetime():
+def test_strip_chart_datetime(num_points: int = 100):
     """Test a strip chart with datetime on the x axis.
     """
-    plt.figure(inspect.currentframe().f_code.co_name)
-    chart = StripChart(datetime=True, xlabel='UTC time')
-    t = np.linspace(1_600_000_000., 1_600_000_900., 100)
-    y = np.sin(np.linspace(0., 10., 100))
-    chart.extend(t, y)
-    chart.plot()
+    t0 = time.time()
+    y = _RNG.random(num_points)
+    for duration in (10, 100, 1000, 10000, 100000):
+        plt.figure(f"{inspect.currentframe().f_code.co_name}_{duration}")
+        chart = EpochStripChart(label="Random data")
+        t = t0 + np.linspace(0., duration, num_points)
+        chart.extend(t, y)
+        chart.plot()
 
 
 if __name__ == '__main__':